created functions to handle floats

+++ srcs/get_values.c
@@ -6,7 +6,7 @@
 /*   By: thuynh <thuynh@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/04 21:05:12 by thuynh            #+#    #+#             */
-/*   Updated: 2017/08/15 17:46:30 by thuynh           ###   ########.fr       */
+/*   Updated: 2017/08/16 10:10:23 by thuynh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,5 +28,5 @@ void	get_values(t_print **list, va_list ap)
 	else if (F == 'p')
 		VAL_ADDR = va_arg(ap, void *);
 	else if (F == 'f' || F == 'F')
-		VAL_DBL = va_arg(ap, long double);
+		VAL_DBL = va_arg(ap, double);
 }
